[DIAR-98] refactor: 파일 위치 변경
